fix(web): tighten numeric field semantics and command help
Clearing the Douyu retry count stores `null` rather than dropping the key, so
it keeps overriding a legacy flat template value the way the sibling CDN field
and the Bilibili quality selector already do.

Retry counts, delays, thumbnail dimensions, compression level and CRF are whole
numbers on the backend, so their schemas reject a fraction and their inputs step
by one; genuinely fractional limits are untouched.

The execute step keeps a description for its command box, with the placeholder
legend rendered after it, so the box's accessible description still resolves.

+++ rust-srec/frontend/src/api/schemas/platform-configs.ts
@@ -50,7 +50,7 @@ export const DouyuConfigSchema = z
     disable_interactive_game: z.boolean().nullable().optional(),
     only_audio: z.boolean().nullable().optional(),
     rate: z.number().nullable().optional(),
-    request_retries: z.number().nullable().optional(),
+    request_retries: z.number().int().nullable().optional(),
     end_stream_on_danmu_stream_closed: z.boolean().nullable().optional(),
   })
   .strict();
